Organize classes into subpackages som.task and som.command
As the number of classes increase, a single package becomes clustered
and hard to navigate.

Let's divide classes into logical subpackages:
* som.task: Task-related classes
* som.command: All command classes

This is done to improve code organization, and each package has a
clear responsibility.
