+++ src/main/java/som/Som.java
@@ -1,5 +1,6 @@
-import java.util.List;
-import java.time.LocalDate;
+package som;
+
+import som.command.Command;
 
 /**
  * The main chatbot class that handles user interaction, task management, and persistence.
