+++ src/main/java/som/Ui.java
@@ -1,5 +1,10 @@
+package som;
+
+import som.task.Task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+
 import java.util.Scanner;
 import java.util.List;
