+++ src/main/java/som/command/Command.java
@@ -1,3 +1,9 @@
+package som.command;
+
+import som.SomException;
+import som.TaskList;
+import som.Ui;
+
 public abstract class Command {
 
     public abstract void execute(TaskList tasks, Ui ui) throws SomException;
