+++ src/main/java/som/command/FindCommand.java
@@ -1,3 +1,9 @@
+package som.command;
+
+import som.SomException;
+import som.TaskList;
+import som.Ui;
+
 public class FindCommand extends Command {
     private String dateStr;
 
@@ -7,7 +13,7 @@ public class FindCommand extends Command {
 
     @Override
     public void execute(TaskList tasks, Ui ui) throws SomException {
-        ui.showTasksOnDate(tasks.getAllTasks(), java.time.LocalDate.parse(dateStr));
+        ui.showTasksOnDate(tasks.findTasks(dateStr), java.time.LocalDate.parse(dateStr));
     }
 
     @Override
