+++ src/main/java/som/command/HelpCommand.java
@@ -1,8 +1,14 @@
-public class ListCommand extends Command {
+package som.command;
+
+import som.SomException;
+import som.TaskList;
+import som.Ui;
+
+public class HelpCommand extends Command {
 
     @Override
     public void execute(TaskList tasks, Ui ui) throws SomException {
-        ui.showList(tasks.getAllTasks());
+        ui.showHelp();
     }
 
     @Override
