+++ src/main/java/som/command/MarkCommand.java
@@ -1,14 +1,21 @@
-public class UnmarkCommand extends Command {
+package som.command;
+
+import som.SomException;
+import som.Storage;
+import som.TaskList;
+import som.Ui;
+
+public class MarkCommand extends Command {
     private int index;
 
-    public UnmarkCommand(int index) {
+    public MarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui) throws SomException {
-        tasks.unmark(index);
-        ui.showUnmarked(tasks.get(index));
+        tasks.mark(index);
+        ui.showMarked(tasks.get(index));
         Storage.save(tasks.getAllTasks());
     }
