+++ src/main/java/som/command/UnmarkCommand.java
@@ -1,14 +1,21 @@
-public class MarkCommand extends Command {
+package som.command;
+
+import som.SomException;
+import som.Storage;
+import som.TaskList;
+import som.Ui;
+
+public class UnmarkCommand extends Command {
     private int index;
 
-    public MarkCommand(int index) {
+    public UnmarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui) throws SomException {
-        tasks.mark(index);
-        ui.showMarked(tasks.get(index));
+        tasks.unmark(index);
+        ui.showUnmarked(tasks.get(index));
         Storage.save(tasks.getAllTasks());
     }
