+++ src/main/java/som/task/Event.java
@@ -1,3 +1,7 @@
+package som.task;
+
+import som.SomException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
