+++ src/main/java/som/task/Task.java
@@ -1,3 +1,5 @@
+package som.task;
+
 /**
  * Represents a generic bask.
  * <p>
